feat: obtain author names from landing page if not found in JSON data
This should fix the "nan" entries that have recently cropped up in the FedSanFranciscoScraper class. The API data has been more error-ridden lately, and the publication_authors field is empty for most new entries. The landing page more reliably has author names, which is now the fallback solution if the JSON data is missing.